scr/requirements.txt
rrequests==2.4.0
BeautifulSoup==3.2.2.tar.gz
logging==0.4.9.6